Move sport modules to a sub-directory
