+++ common/app/assets/javascripts/bootstraps/article.js
@@ -3,10 +3,10 @@ define([
     "modules/autoupdate",
     "modules/live-filter",
     "modules/live-summary",
-    "modules/matchnav",
+    "modules/sport/football/matchnav",
     "modules/analytics/reading",
     "modules/discussion/loader",
-    "modules/cricket",
+    "modules/sport/cricket",
     "modules/experiments/live-blog-show-more",
     "modules/notification-counter",
     "modules/detect",
