+++ common/app/assets/javascripts/modules/sport/cricket.js
