+++ common/app/assets/javascripts/modules/sport/football/tables.js
